Parse markdown-skill frontmatter with the shared YAML parser
parseMarkdownSkill hand-rolled the frontmatter fence and name extraction with
regexes, which breaks on multiline YAML values and quoting. Move the canonical
parseFrontmatter from base into runtime-common (base re-exports it) so prompt
assembly shares the same parser the indexer uses; malformed frontmatter
degrades to raw content instead of crashing prompt assembly.
